fix auto-inserted `photos` shortcode not re-rendering in server mode
This happens because after an initial render, the value used to
track if the shortcode was called persists. Same happens with
`.HasShortcode` (as I later learned), but in that case it's
intentional[^1].

You could reset the manual tracking after checking, but the
'proper' solution is to extract it into a partial anyway and
call said partial from the shortcode.

[^1]: https://redirect.github.com/gohugoio/hugo/issues/6703#issuecomment-1140522919